Codebase: cleaning folder structure
