+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { HeroSection } from './components/HeroSection'
+import { HeroSection } from './components/Hero/HeroSection.component'
 
 function App() {
   return (
