+++ src/components/Hero/Amimation/Animation3D.component.tsx
@@ -1,4 +1,4 @@
-import { useThreeAnimation } from "../../hooks/useThreeAnimation";
+import { useThreeAnimation } from "../../../Hooks/useThreeAnimation.hook";
 
 
 export const Animation3D = () => {
